pyproject.toml
[project]
name = "drf-cookiecutter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-admin-sortable2>=2.2.6",
]
